project/requirements-test.txt
-r requirements.txt
model-mommy
pytest
pytest-cov
pytest-django
requests-mock
ipython